+++ GeneralsMD/Code/GameEngine/Source/GameNetwork/ConnectionManager.cpp
@@ -39,6 +39,7 @@
 #include "Common/Recorder.h"
 
 #include "GameClient/Diplomacy.h"
+#include "GameClient/PlayerColorScheme.h"
 #include "GameClient/GameText.h"
 #include "GameClient/MessageBox.h"
 #include "GameNetwork/ConnectionManager.h"
@@ -706,7 +707,7 @@ void ConnectionManager::processChat(NetChatCommandMsg *msg)
 	if ( ((1<<m_localSlot) & msg->getPlayerMask() ) && canSeeChat  )
 	{
 		RGBColor rgb;
-		rgb.setFromInt(player->getPlayerColor());
+		rgb.setFromInt(clientPlayerColor(player));
 		TheInGameUI->messageColor(&rgb, UnicodeString(L"%ls"), unitext.str());
 
 		// feedback for received chat messages in-game